# apps/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/projects": {
      "post": {
        "description": "Creates a new push notification project with auto-generated API key and configuration.",
        "operationId": "Projects_Create",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "64f1a2b3c4d5e6f7a8b9c0d1",
                    "name": "My Mobile App",
                    "description": "Push notifications for my mobile application",
                    "apiKey": "pns_1234567890abcdef",
                    "settings": {
                      "fcm": {
                        "enabled": true
                      },
                      "apns": {
                        "enabled": false
                      },
                      "webPush": {
                        "enabled": true
                      }
                    },
                    "createdAt": "2023-12-07T10:00:00.000Z",
                    "updatedAt": "2023-12-07T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Create a new project",
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "description": "Returns the project associated with the provided API key.",
        "operationId": "Projects_GetCurrent",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "_id": "64f1a2b3c4d5e6f7a8b9c0d1",
                      "name": "My Mobile App",
                      "description": "Push notifications for my mobile application",
                      "settings": {
                        "fcm": {
                          "enabled": true,
                          "serviceAccountKey": "[REDACTED]"
                        },
                        "apns": {
                          "enabled": false
                        },
                        "webPush": {
                          "enabled": true
                        }
                      },
                      "stats": {
                        "totalDevices": 1250,
                        "totalNotifications": 5420,
                        "deliveryRate": 0.954
                      },
                      "createdAt": "2023-12-07T10:00:00.000Z",
                      "updatedAt": "2023-12-07T15:30:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Get authenticated project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/v1/projects/{id}": {
      "get": {
        "description": "Retrieves a specific project by ID. Users can only access their own project.",
        "operationId": "findOne",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "example": "64f1a2b3c4d5e6f7a8b9c0d1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project found and returned"
          },
          "403": {
            "description": "Access denied - can only access own project",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Access denied to this project",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Get project by ID",
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "description": "Updates project settings and configuration. Users can only update their own project.",
        "operationId": "update",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID to update",
            "schema": {
              "example": "64f1a2b3c4d5e6f7a8b9c0d1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully"
          },
          "403": {
            "description": "Access denied - can only update own project"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Update project",
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "description": "Permanently deletes a project and all associated data. This action cannot be undone.",
        "operationId": "remove",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID to delete",
            "schema": {
              "example": "64f1a2b3c4d5e6f7a8b9c0d1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project deleted successfully"
          },
          "403": {
            "description": "Access denied - can only delete own project"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Delete project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/v1/projects/{id}/regenerate-api-key": {
      "post": {
        "description": "Generates a new API key for the project. The old API key will become invalid immediately.",
        "operationId": "Projects_RotateApiKey",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "example": "64f1a2b3c4d5e6f7a8b9c0d1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key regenerated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "64f1a2b3c4d5e6f7a8b9c0d1",
                    "name": "My Mobile App",
                    "apiKey": "pns_new1234567890abcdef",
                    "message": "API key regenerated successfully. Please update your applications with the new key.",
                    "regeneratedAt": "2023-12-07T16:00:00.000Z"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access denied - can only regenerate API key for own project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Regenerate API key",
        "tags": [
          "Projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Push Notification Service API",
    "description": "\n      Enterprise-grade Push Notification Service supporting iOS, Android, and Web platforms.\n      \n      ## Features\n      - 🚀 Multi-platform push notifications (FCM, APNs, Web Push)\n      - 📱 Advanced device management and segmentation\n      - 📊 Real-time analytics and delivery tracking\n      - 🏷️ Template-based notifications with variable substitution\n      - ⚡ High-performance queue system with BullMQ\n      - 🔒 Enterprise security with API key authentication\n      - 🛡️ Rate limiting for API protection\n      - 📈 99.9% uptime reliability target\n      \n      ## Authentication\n      All endpoints require an API key in the `X-API-Key` header.\n      \n      ## Rate Limiting\n      API endpoints are rate-limited to ensure fair usage:\n      - **High Frequency**: 100 requests/minute (notifications, events)\n      - **Medium Frequency**: 300 requests/minute (device management)\n      - **Low Frequency**: 1000 requests/hour (analytics, admin)\n      \n      Rate limit headers are included in all responses:\n      - `X-RateLimit-Limit`: Maximum requests allowed\n      - `X-RateLimit-Remaining`: Requests remaining in current window\n      - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Push Notification Service Team",
      "url": "https://github.com/your-repo/push-notification-service",
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for project authentication. Get your API key from the project dashboard."
      }
    },
    "schemas": {
      "CreateProjectDto": {
        "type": "object",
        "properties": {}
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "object",
            "example": "Validation failed"
          },
          "timestamp": {
            "type": "string",
            "example": "2025-01-01T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/projects/123/notifications/send"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}